bibdocfile: Types of documents as plugins
It is possible to write a plugin that will define a subclass of BibDoc that will be called automatically.
Instead calling the constructor of BibDoc class, a factory static function create_instance should be used
(It takes the same arguments as the constructor).
A sample plugin defining documents with text content has been separated